+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to **pipecat** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed an issue in `AzureTTSService` that was causing interruptions to take
+  longer. Azure Speech Synthesizer returns big chunks of audio. Currently, we
+  don't have a way to stop a long `AudioRawFrame`, so we have to chunk longer
+  audio into small chunks.
+
 ## [0.0.29] - 2024-06-07
 
 ### Added